+++ src/main/java/com/almostreliable/summoningrituals/compat/AltarObservable.java
@@ -1,9 +1,8 @@
 package com.almostreliable.summoningrituals.compat;
 
+import com.almostreliable.summoningrituals.altar.AltarBlockEntity;
 import com.almostreliable.summoningrituals.recipe.container.RecipeInfo;
 
-import net.minecraft.core.BlockPos;
-import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
 
 import org.jetbrains.annotations.Nullable;
@@ -15,9 +14,9 @@ public class AltarObservable {
 
     private final List<Observer> observers = new ArrayList<>();
 
-    public boolean invoke(ServerLevel level, BlockPos pos, RecipeInfo recipeInfo, @Nullable ServerPlayer player) {
+    public boolean invoke(AltarBlockEntity altar, RecipeInfo recipeInfo, @Nullable ServerPlayer player) {
         for (var o : observers) {
-            if (o.run(level, pos, recipeInfo, player)) return false;
+            if (o.run(altar, recipeInfo, player)) return false;
         }
         return true;
     }
@@ -29,6 +28,6 @@ public class AltarObservable {
     @FunctionalInterface
     public interface Observer {
 
-        boolean run(ServerLevel level, BlockPos pos, RecipeInfo recipeInfo, @Nullable ServerPlayer player);
+        boolean run(AltarBlockEntity altar, RecipeInfo recipeInfo, @Nullable ServerPlayer player);
     }
 }